Update comments Fixed remove the conflict packages if package has not install

+++ include/public.sh
@@ -757,13 +757,13 @@ firewall_set(){
 remove_packages(){
     log "Info" "Starting remove the conflict packages..."
     if check_sys packageManager apt; then
-        apt-get -y remove --purge apache2 apache2-* &> /dev/null
-        apt-get -y remove --purge mysql-client mysql-server mysql-common libmysqlclient18 &> /dev/null
-        apt-get -y remove --purge php5 php5-* &> /dev/null
+        [ "${apache}" != "do_not_install" ] && apt-get -y remove --purge apache2 apache2-* &> /dev/null
+        [ "${mysql}" != "do_not_install" ] && apt-get -y remove --purge mysql-client mysql-server mysql-common libmysqlclient18 &> /dev/null
+        [ "${php}" != "do_not_install" ] && apt-get -y remove --purge php5 php5-* php7.0 php7.0-* php7.1 php7.1-* php7.2 php7.2-* php7.3 php7.3-* &> /dev/null
     elif check_sys packageManager yum; then
-        yum -y remove httpd-* &> /dev/null
-        yum -y remove mysql-* &> /dev/null
-        yum -y remove php-* libzip-devel libzip &> /dev/null
+        [ "${apache}" != "do_not_install" ] && yum -y remove httpd-* &> /dev/null
+        [ "${mysql}" != "do_not_install" ] && yum -y remove mysql-* &> /dev/null
+        [ "${php}" != "do_not_install" ] && yum -y remove php-* libzip-devel libzip &> /dev/null
     fi
     log "Info" "Remove the conflict packages completed..."
 }
